agentic_data_platform/requirements.txt
# AI-Ready Agentic Data Platform - Python Dependencies
# Install with: pip install -r requirements.txt

# ============================================================
# Core Dependencies
# ============================================================
python-dateutil>=2.8.2
pandas>=2.0.0
numpy>=1.24.0

# ============================================================
# Kafka
# ============================================================
kafka-python>=2.0.2

# ============================================================
# Spark & Delta Lake
# ============================================================
pyspark>=3.5.0
delta-spark>=3.1.0
deltalake>=0.14.0  # For Python-native Delta Lake reading

# ============================================================
# Machine Learning
# ============================================================
scikit-learn>=1.3.0
joblib>=1.3.0

# ============================================================
# LLM & Agent
# ============================================================
langchain>=0.1.0
langchain-community>=0.0.10
langchain-ollama>=0.1.0
ollama>=0.1.0
faiss-cpu>=1.7.4

# ============================================================
# Airflow
# ============================================================
apache-airflow>=2.8.0

# ============================================================
# Utilities
# ============================================================
pyarrow>=14.0.0
requests>=2.31.0
tqdm>=4.66.0

# ============================================================
# Dashboard
# ============================================================
streamlit>=1.30.0
plotly>=5.18.0

# ============================================================
# Development (Optional)
# ============================================================
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
